backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
PyMuPDF>=1.23.0
pillow>=10.0.0
httpx>=0.25.0
python-multipart>=0.0.6
PyPDF2==3.0.1
pydantic==2.4.2
colorama>=0.4.6
python-dotenv==1.0.0 